+++ zanata-model/src/main/java/org/zanata/model/HTextFlowTarget.java
@@ -121,7 +121,7 @@ public class HTextFlowTarget extends ModelEntityBase implements HasContents,
     @Getter
     private String revisionComment;
 
-    private boolean revisionCommentChanged = false;
+    private boolean revisionCommentSet = false;
 
     // Only for internal use (persistence transient)
     @Setter(AccessLevel.PRIVATE)
@@ -129,12 +129,12 @@ public class HTextFlowTarget extends ModelEntityBase implements HasContents,
 
     public void setRevisionComment(String revisionComment) {
         this.revisionComment = revisionComment;
-        revisionCommentChanged = true;
+        revisionCommentSet = true;
     }
 
     @Transient
-    boolean isRevisionCommentChanged() {
-        return revisionCommentChanged;
+    boolean isRevisionCommentSet() {
+        return revisionCommentSet;
     }
 
     // Only for internal use (persistence transient)
@@ -427,7 +427,7 @@ public class HTextFlowTarget extends ModelEntityBase implements HasContents,
             // insert history if this has changed from its initial state
             if (tft.initialState != null && tft.initialState.hasChanged(tft)) {
                 tft.getHistory().put(tft.oldVersionNum, tft.initialState);
-                if (!tft.isRevisionCommentChanged()) {
+                if (!tft.isRevisionCommentSet()) {
                     tft.setRevisionComment(null);
                 }
             }